backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.6
pandas==2.2.3
python-multipart==0.0.20
python-dotenv==1.0.1
